+++ Servers/routes/file.route.ts
@@ -12,28 +12,12 @@ import {
 } from "../controllers/file.ctrl";
 import authenticateJWT from "../middleware/auth.middleware";
 import authorize from "../middleware/accessControl.middleware";
-const multer = require("multer");
-import * as path from "path";
-import { ALLOWED_MIME_TYPES } from "../utils/validations/fileManagerValidation.utils";
+import { createMemoryUpload } from "../utils/upload.utils";
 
 // Security: the previous config had a typo'd `Storage` key, so multer ran
 // with zero limits and no type validation (memory-exhaustion DoS vector).
-const upload = multer({
-  storage: multer.memoryStorage(),
-  limits: {
-    fileSize: 30 * 1024 * 1024, // 30MB per file
-    files: 10,
-  },
-  fileFilter: (_req: Express.Request, file: Express.Multer.File, cb: any) => {
-    const allowedExts = ALLOWED_MIME_TYPES[file.mimetype as keyof typeof ALLOWED_MIME_TYPES];
-    const ext = path.extname(file.originalname).toLowerCase();
-    if (allowedExts && Array.isArray(allowedExts) && allowedExts.includes(ext)) {
-      cb(null, true);
-    } else {
-      cb(new Error("UNSUPPORTED_FILE_TYPE"));
-    }
-  },
-});
+// Use the shared bounded factory (30MB/file, 10 files, MIME allowlist).
+const upload = createMemoryUpload();
 
 const router = express.Router();
